refactor: converted contact form to global component.
