+++ front/reyesandfriends.cl/src/pages/contact/contact.tsx
@@ -1,80 +1,9 @@
-import { useGetContactCategories } from "../../hooks/contact/useGetContactCategories";
-import { useContactFormValidator } from "../../hooks/contact/useContactFormValidator";
-import { useState, useRef } from "react";
-import ContactModal from "./modal/contactModal";
 import { Helmet } from "react-helmet-async";
 import { Contact2 } from "lucide-react";
 import HeroSection from "../../layouts/components/hero-section/hero-section";
-
-function capitalizeWords(str: string) {
-    return str.replace(/\b\w/g, char => char.toUpperCase()).replace(/\B\w/g, char => char.toLowerCase());
-}
+import ContactForm from "./components/ContactForm";
 
 const Contact = () => {
-    const { categories, error: categoriesError, loading, refetch } = useGetContactCategories();
-    const { errors, handleSubmit, isSubmitting, finalMessage, setFinalMessage } = useContactFormValidator();
-    const [isFormValid, setIsFormValid] = useState(false);
-    const formRef = useRef<HTMLFormElement>(null);
-    const [cellphone, setCellphone] = useState("");
-    const [firstName, setFirstName] = useState("");
-    const [lastName, setLastName] = useState("");
-
-    const handleCellphoneChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const value = e.target.value;
-        if (/^\d*$/.test(value)) {
-            setCellphone(value);
-        }
-    };
-
-    const handleFirstNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const value = capitalizeWords(e.target.value);
-        setFirstName(value);
-    };
-
-    const handleLastNameChange = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const value = capitalizeWords(e.target.value);
-        setLastName(value);
-    };
-
-    const handleFormSubmit = (e: React.FormEvent<HTMLFormElement>) => {
-        e.preventDefault();
-        const formData = new FormData(e.currentTarget);
-        const data = {
-            first_name: firstName,
-            last_name: lastName,
-            cellphone: cellphone,
-            email: formData.get("email"),
-            category: formData.get("category"),
-            message: formData.get("message"),
-        };
-        handleSubmit(data);
-    };
-
-    const handleInputChange = (e: React.ChangeEvent<HTMLFormElement>) => {
-        const formData = new FormData(e.currentTarget);
-        const isValid = [
-            firstName.trim(),
-            lastName.trim(),
-            cellphone.trim(),
-            formData.get("email"),
-            formData.get("category"),
-            formData.get("message")
-        ].every((value) => {
-            if (typeof value === "string") return value.trim() !== "";
-            return value !== null && value !== undefined;
-        });
-        setIsFormValid(isValid);
-    };
-
-    const handleModalClose = () => {
-        setFinalMessage(null);
-        formRef.current?.reset();
-        setCellphone("");
-        setFirstName("");
-        setLastName("");
-        setIsFormValid(false);
-    };
-
     return (
         <>
             <Helmet>
@@ -87,23 +16,16 @@ const Contact = () => {
                 <meta property="og:description" content="Ponte en contacto con Reyes&Friends para consultas, soporte y más información sobre nuestros servicios." />
                 <meta property="og:image" content="/img/open-graph-images/contact-us.png" />
                 <meta property="og:type" content="website" />
-
                 <meta name="twitter:title" content="Contacto | Reyes&Friends" />
                 <meta name="twitter:description" content="Ponte en contacto con Reyes&Friends para consultas, soporte y más información sobre nuestros servicios." />
                 <meta name="twitter:image" content="/img/open-graph-images/contact-us.png" />
             </Helmet>
-            <ContactModal
-                message={finalMessage}
-                onClose={handleModalClose}
-            />
-
             <HeroSection 
-            icon={Contact2}
-            logoImage="/img/contact/illustration.png"
-            title="Contacto"
-            subtitle="En Reyes&Friends, estamos aquí para ayudarte. Si tienes alguna pregunta o inquietud, no dudes en ponerte en contacto con nosotros."
+                icon={Contact2}
+                logoImage="/img/contact/illustration.png"
+                title="Contacto"
+                subtitle="En Reyes&Friends, estamos aquí para ayudarte. Si tienes alguna pregunta o inquietud, no dudes en ponerte en contacto con nosotros."
             />
-
             <section className="py-16 bg-zinc-900">
                 <div className="container mx-auto px-4 max-w-7xl">
                     <div className="relative mb-12">
@@ -112,128 +34,7 @@ const Contact = () => {
                             <span className="bg-zinc-900 px-4 relative z-10 text-white">Envíanos un mensaje</span>
                         </h2>
                     </div>
-                    <div className="bg-black p-8 rounded-lg shadow-lg">
-                        {(categoriesError || loading) ? (
-                            <div className="flex flex-col items-center justify-center py-12">
-                                <h2 className="text-reyes text-2xl font-bold mb-2">Error al cargar</h2>
-                                <p className="text-white text-lg mb-4 text-center">
-                                    {loading ? "Recargando..." : categoriesError}
-                                </p>
-                                <button
-                                    className="bg-reyes text-white font-bold py-2 px-6 rounded-sm hover:bg-reyes-dark transition-colors"
-                                    onClick={refetch}
-                                    disabled={loading}
-                                >
-                                    {loading ? "Recargando..." : "Reintentar"}
-                                </button>
-                            </div>
-                        ) : (
-                            <form
-                                ref={formRef}
-                                className="grid grid-cols-1 md:grid-cols-2 gap-4"
-                                onSubmit={handleFormSubmit}
-                                onChange={handleInputChange}
-                            >
-                                <div className="col-span-1">
-                                    <label htmlFor="name" className="block text-gray-300 font-bold mb-2">Nombre (requerido)</label>
-                                    <input
-                                        type="text"
-                                        id="first_name"
-                                        name="first_name"
-                                        className="w-full p-3 rounded-sm bg-zinc-800 text-white border border-zinc-700 focus:outline-none focus:ring-2 focus:ring-reyes"
-                                        placeholder="Ingresa tu nombre"
-                                        value={firstName}
-                                        onChange={handleFirstNameChange}
-                                    />
-                                    {errors.name && <p className="text-reyes text-sm">{errors.name}</p>}
-                                </div>
-                                <div className="col-span-1">
-                                    <label htmlFor="last_name" className="block text-gray-300 font-bold mb-2">Apellido (requerido)</label>
-                                    <input
-                                        type="text"
-                                        id="last_name"
-                                        name="last_name"
-                                        className="w-full p-3 rounded-sm bg-zinc-800 text-white border border-zinc-700 focus:outline-none focus:ring-2 focus:ring-reyes"
-                                        placeholder="Ingresa tu apellido"
-                                        value={lastName}
-                                        onChange={handleLastNameChange}
-                                    />
-                                    {errors.last_name && <p className="text-reyes text-sm">{errors.last_name}</p>}
-                                </div>
-                                <div className="col-span-1">
-                                    <label htmlFor="cellphone" className="block text-gray-300 font-bold mb-2">Número de Teléfono (requerido)</label>
-                                    <div className="flex">
-                                        <span className="inline-flex items-center px-3 rounded-l-sm bg-zinc-800 text-white border border-r-0 border-zinc-700 select-none">
-                                            +56
-                                        </span>
-                                        <input
-                                            type="tel"
-                                            id="cellphone"
-                                            name="cellphone"
-                                            maxLength={9}
-                                            className="w-full p-3 rounded-r-sm bg-zinc-800 text-white border border-zinc-700 border-l-0 focus:outline-none focus:ring-2 focus:ring-reyes"
-                                            placeholder="912345678"
-                                            pattern="[0-9]{9}"
-                                            inputMode="numeric"
-                                            value={cellphone}
-                                            onChange={handleCellphoneChange}
-                                        />
-                                    </div>
-                                    {errors.cellphone && <p className="text-reyes text-sm">{errors.cellphone}</p>}
-                                </div>
-                                <div className="col-span-1">
-                                    <label htmlFor="email" className="block text-gray-300 font-bold mb-2">Email (requerido)</label>
-                                    <input
-                                        type="email"
-                                        id="email"
-                                        name="email"
-                                        className="w-full p-3 rounded-sm bg-zinc-800 text-white border border-zinc-700 focus:outline-none focus:ring-2 focus:ring-reyes"
-                                        placeholder="Ingresa tu email"
-                                    />
-                                    {errors.email && <p className="text-reyes text-sm">{errors.email}</p>}
-                                </div>
-                                <div className="md:col-span-2 col-span-1">
-                                    <label htmlFor="category" className="block text-gray-300 font-bold mb-2">Categoría (requerido)</label>
-                                    <select
-                                        id="category"
-                                        name="category"
-                                        className="w-full p-3 rounded-sm bg-zinc-800 text-white border border-zinc-700 focus:outline-none focus:ring-2 focus:ring-reyes"
-                                    >
-                                        <option value="">Selecciona una categoría</option>
-                                        {categories.map((category, index) => (
-                                            <option key={index} value={category.slug}>
-                                                {category.name}
-                                            </option>
-                                        ))}
-                                    </select>
-                                    {errors.category && <p className="text-reyes text-sm">{errors.category}</p>}
-                                </div>
-                                <div className="col-span-1">
-                                    <label htmlFor="message" className="block text-gray-300 font-bold mb-2 md:hidden">Mensaje (requerido)</label>
-                                </div>
-                                <div className="md:col-span-2">
-                                    <label htmlFor="message" className="block text-gray-300 font-bold mb-2 hidden md:block">Mensaje (requerido)</label>
-                                    <textarea
-                                        id="message"
-                                        name="message"
-                                        rows={5}
-                                        className="w-full p-3 rounded-sm bg-zinc-800 text-white border border-zinc-700 focus:outline-none focus:ring-2 focus:ring-reyes"
-                                        placeholder="Escribe tu mensaje aquí..."
-                                    ></textarea>
-                                    {errors.message && <p className="text-reyes text-sm">{errors.message}</p>}
-                                </div>
-                                <div className="md:col-span-2 col-span-1">
-                                    <button
-                                        type="submit"
-                                        className="w-full bg-reyes text-white font-bold py-3 rounded-sm hover:bg-reyes-dark transition-colors disabled:bg-gray-600 disabled:cursor-not-allowed"
-                                        disabled={!isFormValid || isSubmitting}
-                                    >
-                                        {isSubmitting ? "Enviando..." : "Enviar"}
-                                    </button>
-                                </div>
-                            </form>
-                        )}
-                    </div>
+                    <ContactForm />
                 </div>
             </section>
         </>
